Make alloc_clobber_counter of type u16 instead of s64 so that long-running in-process fuzzing jobs would still detect uses of uninitialized memory as small as two bytes

+++ libdislocator.so.c
@@ -66,7 +66,7 @@
 /* Canary & clobber bytes: */
 
 #define ALLOC_CANARY  0xAACCAACC
-int alloc_clobber_counter = 0;
+u16 alloc_clobber_counter = 0;
 
 #define PTR_C(_p) (((u32*)(_p))[-1])
 #define PTR_L(_p) (((u32*)(_p))[-2])
